Рефакторинг: раскладывает по папкам типы и моки
