+++ src/components/app/app.tsx
@@ -1,5 +1,5 @@
 import {BrowserRouter, Route, Routes} from 'react-router-dom';
-import { AppRoute, AuthorizationStatus, OffersType } from '../../const';
+import { AppRoute, AuthorizationStatus } from '../../const';
 import {HelmetProvider} from 'react-helmet-async';
 import MainScreen from '../../pages/main/main-screen';
 import LoginScreen from '../../pages/login/login-screen';
@@ -7,31 +7,17 @@ import FavoritesScreen from '../../pages/favorites/favorites-screen';
 import OfferScreen from '../../pages/offer/offer-screen';
 import NotFoundScreen from '../../pages/not-found-screen/not-found-screen';
 import PrivateRoute from '../private-route/private-route';
+import { OffersCount } from '../../types/CommonTypes.type';
+import { Cities } from '../../types/CommonTypes.type';
+import { Offer } from '../../types/Offer.type';
+import { OffersType } from '../../types/Offers.type';
+
 
 type AppScreenProps = {
-  offersCount: number;
-  cities: string[];
+  offersCount: OffersCount;
+  cities: Cities;
   offers: OffersType[];
-  offer: {
-    images: string[];
-    title: string;
-    isFavorite: boolean;
-    isPremium: boolean;
-    rating: number;
-    type: string;
-    bedrooms: number;
-    maxAdults: number;
-    previewImage: string;
-    price: number;
-    goods: string[];
-    host: {
-      id: number;
-      name: string;
-      isPro: boolean;
-      avatarUrl: string;
-    };
-    id: number;
-  };
+  offer: Offer;
 }
 
 function App({offersCount, cities, offers, offer}: AppScreenProps): JSX.Element {
