+++ src/const.ts
@@ -1,338 +1,6 @@
 export const Setting = {
-  OffersCount: 312
+  OffersCount: 5
 };
-
-export const cities: string[] = [
-  'Paris',
-  'Cologne',
-  'Brussels',
-  'Amsterdam',
-  'Hamburg',
-  'Dusseldorf'
-];
-
-export type OffersType = {
-  city: {
-    name: string;
-  };
-  previewImage: string;
-  images: string[];
-  title: string;
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-  type: string;
-  bedrooms: number;
-  maxAdults: number;
-  price: number;
-  goods: string[];
-  host: {
-    id: number;
-    name: string;
-    isPro: boolean;
-    avatarUrl: string;
-  };
-  description: string;
-  id: number;
-}
-
-export const offerData: OffersType = {
-  city: {
-    name: 'Dusseldorf',
-  },
-  previewImage: 'https://14.react.pages.academy/static/offer/3.jpg',
-  images: [
-    'https://14.react.pages.academy/static/offer/16.jpg',
-    'https://14.react.pages.academy/static/offer/13.jpg',
-    'https://14.react.pages.academy/static/offer/6.jpg',
-    'https://14.react.pages.academy/static/offer/11.jpg',
-    'https://14.react.pages.academy/static/offer/14.jpg',
-    'https://14.react.pages.academy/static/offer/9.jpg',
-    'https://14.react.pages.academy/static/offer/2.jpg',
-    'https://14.react.pages.academy/static/offer/15.jpg',
-    'https://14.react.pages.academy/static/offer/7.jpg',
-    'https://14.react.pages.academy/static/offer/17.jpg',
-    'https://14.react.pages.academy/static/offer/4.jpg',
-    'https://14.react.pages.academy/static/offer/8.jpg',
-    'https://14.react.pages.academy/static/offer/10.jpg',
-    'https://14.react.pages.academy/static/offer/5.jpg'
-  ],
-  title: 'Waterfront with extraordinary view',
-  isFavorite: false,
-  isPremium: true,
-  rating: 4.8,
-  type: 'room',
-  bedrooms: 1,
-  maxAdults: 2,
-  price: 142,
-  goods: [
-    'Laptop friendly workspace',
-    'Breakfast'
-  ],
-  host: {
-    id: 25,
-    name: 'Angelina',
-    isPro: true,
-    avatarUrl: 'img/avatar-angelina.jpg'
-  },
-  description: 'A new spacious villa, one floor. All commodities, jacuzzi and beautiful scenery. Ideal for families or friends.',
-  id: 1
-};
-
-export const offersData: OffersType[] = [
-  {
-    city: {
-      name: 'Dusseldorf',
-    },
-    previewImage: 'https://14.react.pages.academy/static/offer/3.jpg',
-    images: [
-      'https://14.react.pages.academy/static/offer/16.jpg',
-      'https://14.react.pages.academy/static/offer/13.jpg',
-      'https://14.react.pages.academy/static/offer/6.jpg',
-      'https://14.react.pages.academy/static/offer/11.jpg',
-      'https://14.react.pages.academy/static/offer/14.jpg',
-      'https://14.react.pages.academy/static/offer/9.jpg',
-      'https://14.react.pages.academy/static/offer/2.jpg',
-      'https://14.react.pages.academy/static/offer/15.jpg',
-      'https://14.react.pages.academy/static/offer/7.jpg',
-      'https://14.react.pages.academy/static/offer/17.jpg',
-      'https://14.react.pages.academy/static/offer/4.jpg',
-      'https://14.react.pages.academy/static/offer/8.jpg',
-      'https://14.react.pages.academy/static/offer/10.jpg',
-      'https://14.react.pages.academy/static/offer/5.jpg'
-    ],
-    title: 'Waterfront with extraordinary view',
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.8,
-    type: 'room',
-    bedrooms: 1,
-    maxAdults: 2,
-    price: 142,
-    goods: [
-      'Laptop friendly workspace',
-      'Breakfast'
-    ],
-    host: {
-      id: 25,
-      name: 'Angelina',
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg'
-    },
-    description: 'A new spacious villa, one floor. All commodities, jacuzzi and beautiful scenery. Ideal for families or friends.',
-    id: 1
-  },
-  {
-    city: {
-      name: 'Amsterdam',
-    },
-    previewImage: 'https://14.react.pages.academy/static/offer/14.jpg',
-    images: [
-      'https://14.react.pages.academy/static/offer/8.jpg',
-      'https://14.react.pages.academy/static/offer/7.jpg',
-      'https://14.react.pages.academy/static/offer/17.jpg',
-      'https://14.react.pages.academy/static/offer/16.jpg',
-      'https://14.react.pages.academy/static/offer/1.jpg',
-      'https://14.react.pages.academy/static/offer/14.jpg',
-      'https://14.react.pages.academy/static/offer/12.jpg',
-      'https://14.react.pages.academy/static/offer/2.jpg',
-      'https://14.react.pages.academy/static/offer/10.jpg',
-      'https://14.react.pages.academy/static/offer/6.jpg',
-      'https://14.react.pages.academy/static/offer/20.jpg',
-      'https://14.react.pages.academy/static/offer/4.jpg',
-      'https://14.react.pages.academy/static/offer/5.jpg',
-      'https://14.react.pages.academy/static/offer/13.jpg'
-    ],
-    title: 'House in countryside',
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.8,
-    type: 'room',
-    bedrooms: 1,
-    maxAdults: 2,
-    price: 142,
-    goods: [
-      'Laptop friendly workspace',
-      'Breakfast'
-    ],
-    host: {
-      id: 25,
-      name: 'Angelina',
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg'
-    },
-    description: 'This is a place for dreamers to reset, reflect, and create. Designed with a \'slow\' pace in mind, our hope is that you enjoy every part of your stay; from making local coffee by drip in the morning, choosing the perfect record to put on as the sun sets.',
-    id: 2
-  },
-  {
-    city: {
-      name: 'Brussels',
-    },
-    previewImage: 'https://14.react.pages.academy/static/offer/3.jpg',
-    images: [
-      'https://14.react.pages.academy/static/offer/9.jpg',
-      'https://14.react.pages.academy/static/offer/3.jpg',
-      'https://14.react.pages.academy/static/offer/5.jpg',
-      'https://14.react.pages.academy/static/offer/18.jpg',
-      'https://14.react.pages.academy/static/offer/8.jpg',
-      'https://14.react.pages.academy/static/offer/10.jpg',
-      'https://14.react.pages.academy/static/offer/1.jpg',
-      'https://14.react.pages.academy/static/offer/19.jpg',
-      'https://14.react.pages.academy/static/offer/16.jpg',
-      'https://14.react.pages.academy/static/offer/7.jpg',
-      'https://14.react.pages.academy/static/offer/6.jpg',
-      'https://14.react.pages.academy/static/offer/14.jpg',
-      'https://14.react.pages.academy/static/offer/4.jpg',
-      'https://14.react.pages.academy/static/offer/12.jpg'
-    ],
-    title: 'Loft Studio in the Central Area',
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.8,
-    type: 'room',
-    bedrooms: 1,
-    maxAdults: 2,
-    price: 142,
-    goods: [
-      'Laptop friendly workspace',
-      'Breakfast'
-    ],
-    host: {
-      id: 25,
-      name: 'Angelina',
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg'
-    },
-    description: 'I am happy to welcome you to my apartment in the city center! Three words: location, cosy and chic!',
-    id: 3
-  },
-  {
-    city: {
-      name: 'Amsterdam',
-    },
-    previewImage: 'https://14.react.pages.academy/static/offer/6.jpg',
-    images: [
-      'https://14.react.pages.academy/static/offer/15.jpg',
-      'https://14.react.pages.academy/static/offer/18.jpg',
-      'https://14.react.pages.academy/static/offer/3.jpg',
-      'https://14.react.pages.academy/static/offer/6.jpg',
-      'https://14.react.pages.academy/static/offer/19.jpg',
-      'https://14.react.pages.academy/static/offer/2.jpg',
-      'https://14.react.pages.academy/static/offer/17.jpg',
-      'https://14.react.pages.academy/static/offer/10.jpg',
-      'https://14.react.pages.academy/static/offer/1.jpg',
-      'https://14.react.pages.academy/static/offer/4.jpg',
-      'https://14.react.pages.academy/static/offer/16.jpg',
-      'https://14.react.pages.academy/static/offer/20.jpg',
-      'https://14.react.pages.academy/static/offer/13.jpg',
-      'https://14.react.pages.academy/static/offer/5.jpg'
-    ],
-    title: 'Nice, cozy, warm big bed apartment',
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.8,
-    type: 'room',
-    bedrooms: 1,
-    maxAdults: 2,
-    price: 142,
-    goods: [
-      'Laptop friendly workspace',
-      'Breakfast'
-    ],
-    host: {
-      id: 25,
-      name: 'Angelina',
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg'
-    },
-    description: 'Design interior in most sympathetic area! Complitely renovated, well-equipped, cosy studio in idyllic, over 100 years old wooden house. Calm street, fast connection to center and airport.',
-    id: 4
-  },
-  {
-    city: {
-      name: 'Hamburg',
-    },
-    previewImage: 'https://14.react.pages.academy/static/offer/20.jpg',
-    images: [
-      'https://14.react.pages.academy/static/offer/20.jpg',
-      'https://14.react.pages.academy/static/offer/18.jpg',
-      'https://14.react.pages.academy/static/offer/13.jpg',
-      'https://14.react.pages.academy/static/offer/8.jpg',
-      'https://14.react.pages.academy/static/offer/16.jpg',
-      'https://14.react.pages.academy/static/offer/12.jpg',
-      'https://14.react.pages.academy/static/offer/6.jpg',
-      'https://14.react.pages.academy/static/offer/7.jpg',
-      'https://14.react.pages.academy/static/offer/9.jpg',
-      'https://14.react.pages.academy/static/offer/3.jpg',
-      'https://14.react.pages.academy/static/offer/17.jpg',
-      'https://14.react.pages.academy/static/offer/2.jpg',
-      'https://14.react.pages.academy/static/offer/15.jpg',
-      'https://14.react.pages.academy/static/offer/1.jpg'
-    ],
-    title: 'The house among olive ',
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.8,
-    type: 'room',
-    bedrooms: 1,
-    maxAdults: 2,
-    price: 142,
-    goods: [
-      'Laptop friendly workspace',
-      'Breakfast'
-    ],
-    host: {
-      id: 25,
-      name: 'Angelina',
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg'
-    },
-    description: 'Peaceful studio in the most wanted area in town. Quiet house Near of everything. Completely renovated. Lovely neighbourhood, lot of trendy shops, restaurants and bars in a walking distance.',
-    id: 5
-  },
-  {
-    city: {
-      name: 'Dusseldorf',
-    },
-    previewImage: 'https://14.react.pages.academy/static/offer/13.jpg',
-    images: [
-      'https://14.react.pages.academy/static/offer/15.jpg',
-      'https://14.react.pages.academy/static/offer/20.jpg',
-      'https://14.react.pages.academy/static/offer/18.jpg',
-      'https://14.react.pages.academy/static/offer/3.jpg',
-      'https://14.react.pages.academy/static/offer/16.jpg',
-      'https://14.react.pages.academy/static/offer/10.jpg',
-      'https://14.react.pages.academy/static/offer/11.jpg',
-      'https://14.react.pages.academy/static/offer/14.jpg',
-      'https://14.react.pages.academy/static/offer/5.jpg',
-      'https://14.react.pages.academy/static/offer/7.jpg',
-      'https://14.react.pages.academy/static/offer/1.jpg',
-      'https://14.react.pages.academy/static/offer/12.jpg',
-      'https://14.react.pages.academy/static/offer/2.jpg',
-      'https://14.react.pages.academy/static/offer/13.jpg'
-    ],
-    title: 'Waterfront with extraordinary view',
-    isFavorite: false,
-    isPremium: false,
-    rating: 4.8,
-    type: 'room',
-    bedrooms: 1,
-    maxAdults: 2,
-    price: 142,
-    goods: [
-      'Laptop friendly workspace',
-      'Breakfast'
-    ],
-    host: {
-      id: 25,
-      name: 'Angelina',
-      isPro: true,
-      avatarUrl: 'img/avatar-angelina.jpg'
-    },
-    description: 'A new spacious villa, one floor. All commodities, jacuzzi and beautiful scenery. Ideal for families or friends.',
-    id: 6
-  }
-];
 
 export enum AppRoute {
   Main = '/',
