+++ src/index.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
-import { Setting, cities, offersData, offerData } from './const';
+import { Setting } from './const';
+import { cities } from './mock/locations-data';
+import { offerData } from './mock/offers-data';
+import { offersData } from './mock/offers-data';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
