+++ src/types/CommonTypes.type.tsx
@@ -1,0 +1,3 @@
+export type OffersCount = number;
+
+export type Cities = string[];
