+++ src/types/Host.type.tsx
@@ -1,0 +1,6 @@
+export type Host = {
+    id: number;
+    name: string;
+    isPro: boolean;
+    avatarUrl: string;
+};
